% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{cum2incr}
\alias{cum2incr}
\title{Convert cumulative triangle to incremental one.}
\usage{
cum2incr(triangle)
}
\arguments{
\item{triangle}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
