% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.glmSim}
\alias{.glmSim}
\title{Leading NA}
\usage{
.glmSim(triangle, n_boot, config, type)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if 
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
