% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glm_boot}
\alias{glm_boot}
\title{Simulate Poisson GLM reserve.}
\usage{
glm_boot(triangle, n_boot)
}
\arguments{
\item{triangle}{Incremental claims triangle}
}
\description{
Simulate Poisson GLM reserve.
}
