% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glm_sim}
\alias{glm_sim}
\title{Simulate Poisson GLM reserve for different perturbed and excluded points.}
\usage{
glm_sim(triangle, n_boot, config, type)
}
\arguments{
\item{triangle}{Incremental claims triangle}

\item{n_boot}{Number of bootstrap simulations}

\item{sim_type}{Simulation type: `"single"` (the default), `"calender"`, or `"origin"`}

\item{factor}{Perturbation factor}

\item{boot_type}{Type of bootstrap: `"parametric"`, `"residuals"`, or `"pairs"`}

\item{proc_dist}{Distribution of process error: `"normal"` or `"gamma"`}

\item{cond}{Specified whether the bootstrap should be conditional or unconditional. Default is `TRUE`.}
}
\description{
Simulate Poisson GLM reserve for different perturbed and excluded points.
}
\details{
The simulation configuration inputs `sim_type`, `factor`, `boot_type`, `proc_dist` and `conditional` can be
either strings or character vectors. In the latter case, the simulation is computed for all feasible combinations.
}
