% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{glmConfig}
\alias{glmConfig}
\title{Generate configuration for GLM bootstrap simulations.}
\usage{
glmConfig(ndev, factors, type = "single")
}
\arguments{
\item{ndev}{DESCRIPTION.}

\item{factors}{DESCRIPTION.}

\item{type}{DESCRIPTION.}

\item{resids.type}{DESCRIPTION.}

\item{boot.type}{DESCRIPTION.}

\item{dist}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
