% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{incr2cum}
\alias{incr2cum}
\title{Convert incremental triangle to cumulative one.}
\usage{
incr2cum(triangle)
}
\arguments{
\item{triangle}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
