% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{mackConfig}
\alias{mackConfig}
\title{Generate configuration for Mack bootstrap simulations.}
\usage{
mackConfig(
  ndev,
  factors,
  resids.type = c("raw", "scaled", "parametric"),
  boot.type = c("conditional", "unconditional"),
  dist = c("normal", "gamma"),
  type = "single"
)
}
\arguments{
\item{ndev}{DESCRIPTION.}

\item{factors}{DESCRIPTION.}

\item{resids.type}{DESCRIPTION.}

\item{boot.type}{DESCRIPTION.}

\item{dist}{DESCRIPTION.}

\item{type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
