% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.r
\name{mackSim}
\alias{mackSim}
\title{Bootstrap simulation based on Mack's model.}
\usage{
mackSim(triangle, nboot, config, type)
}
\arguments{
\item{triangle}{DESCRIPTION.}

\item{nboot}{DESCRIPTION.}

\item{config}{DESCRIPTION.}

\item{type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
