% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternBreak-package.R
\docType{package}
\name{patternBreak-package}
\alias{patternBreak}
\alias{patternBreak-package}
\title{patternBreak: Bootstrap Simulations for Claims Reserving}
\description{
More about what it does (maybe more than one line)
}
\keyword{internal}
