% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternBreak-package.R
\docType{package}
\name{patternBreak}
\alias{patternBreak}
\alias{patternBreak-package}
\title{PatternBreak}
\description{
Description of your package
}
\author{
Othman El Hammouchi <youremail>
}
\keyword{internal}
