% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{glmBoot}
\alias{glmBoot}
\title{Compute bootstrap reserve based on GLM model.}
\usage{
glmBoot(triangle, nboot)
}
\arguments{
\item{triangle}{Run-off triangle.}

\item{nboot}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
