% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mackBoot}
\alias{mackBoot}
\title{Leading NA}
\usage{
mackBoot(triangle, n_boot, resids_type, boot_type, dist)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if 
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
