% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mackBoot}
\alias{mackBoot}
\title{Simulate Mack CL reserve.}
\usage{
mackBoot(triangle, n_boot, boot_type, process_dist, conditional, resids_type)
}
\arguments{
\item{triangle}{Cumulative claims triangle}
}
\description{
Simulate Mack CL reserve.
}
