% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.r
\name{mackBoot}
\alias{mackBoot}
\title{Compute bootstrap reserve based on Mack's model.}
\usage{
mackBoot(triangle, nboot, resids.type, boot.type, dist)
}
\arguments{
\item{triangle}{Run-off triangle.}

\item{nboot}{Number of bootstrap samples.}

\item{resids.type}{Type of residuals.}

\item{boot.type}{Type of bootstrap.}

\item{dist}{Distribution.}
}
\value{
A vector of bootstrap reserve estimates.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
