% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mack.r
\name{mackBoot}
\alias{mackBoot}
\title{Bootstrap simulation based on Mack's model.}
\usage{
mackBoot(triangle, nboot, resids.type, boot.type, dist)
}
\arguments{
\item{triangle}{DESCRIPTION.}

\item{nboot}{DESCRIPTION.}

\item{config}{DESCRIPTION.}

\item{type}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
