% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{mackConfig}
\alias{mackConfig}
\title{mackConfig}
\usage{
mackConfig(
  ndev,
  outlier.points = NULL,
  excl.points = NULL,
  factors,
  resids.type = c("raw", "scaled", "parametric"),
  boot.type = c("conditional", "unconditional"),
  dist = c("normal", "gamma")
)
}
\arguments{
\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
Generate configuration
}
\examples{
# ADD_EXAMPLES_HERE
}
