% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mackSim}
\alias{mackSim}
\title{Simulate Mack CL reserve for different perturbed and excluded points.}
\usage{
mackSim(
  triangle,
  sim_type,
  n_boot,
  factors,
  boot_type,
  process_dist,
  conditional = TRUE,
  resids_type = "none",
  show_progress = TRUE
)
}
\arguments{
\item{triangle}{Cumulative claims triangle}

\item{sim_type}{Simulation type: `"single"` (the default), `"calender"`, or `"origin"`}

\item{n_boot}{Number of bootstrap simulations}

\item{factors}{Vector of perturbation factors}

\item{boot_type}{Type of bootstrap: `"parametric"`, `"non-parametric"`, or `"pairs"`}

\item{process_dist}{Distribution of process error: `"normal"` or `"gamma"`}

\item{conditional}{Specified whether the bootstrap should be conditional or unconditional. Default is `TRUE`.}
}
\description{
Simulate Mack CL reserve for different perturbed and excluded points.
}
