% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mackSim}
\alias{mackSim}
\title{Simulate Mack CL reserve for different perturbed and excluded points.}
\usage{
mackSim(triangle, sim_type, n_boot, factors, boot_types, proc_dists, conds,
  resids_type = NULL, show_progress = TRUE)
}
\arguments{
\item{triangle}{Cumulative claims triangle}

\item{sim_type}{Simulation type: `"single"` (the default), `"calender"`, or `"origin"`}

\item{n_boot}{Number of bootstrap simulations}

\item{factor}{Perturbation factor}

\item{boot_type}{Type of bootstrap: `"parametric"`, `"residuals"`, or `"pairs"`}

\item{proc_dist}{Distribution of process error: `"normal"` or `"gamma"`}

\item{cond}{Specified whether the bootstrap should be conditional or unconditional. Default is `TRUE`.}
}
\description{
Simulate Mack CL reserve for different perturbed and excluded points.
}
\details{
The simulation configuration inputs `sim_type`, `factor`, `boot_type`, `proc_dist` and `conditional` can be
either strings or character vectors. In the latter case, the simulation is computed for all feasible combinations.
}
